backend/requirements.txt
accelerate==0.27.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.2.0
async-timeout==4.0.3
attrs==22.2.0
autopep8==1.6.0
babel==2.18.0
beautifulsoup4==4.11.1
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
blis==0.7.11
bottle==0.12.21
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==5.5.1
catalogue==2.0.10
celery==5.6.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.3
comtypes==1.1.11
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.13
datasets==2.17.0
debugpy==1.8.20
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
distlib==0.4.0
distro==1.9.0
Django==5.0.1
django-axes==8.3.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
docker==7.0.0
docopt==0.6.2
Eel==0.14.0
emoji==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
enum34==1.1.10
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.109.0
fastjsonschema==2.21.2
filelock==3.20.0
Flask==2.1.2
flask-cors==6.0.2
fonttools==4.61.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.3.3
fsspec==2023.10.0
future==0.18.2
gensim==4.3.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
googletrans==4.0.2
greenlet==1.1.2
groq==1.0.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hiredis==3.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
hugchat==0.4.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.3
importlib-metadata==4.11.4
importlib_resources==6.5.2
ipykernel==6.28.0
ipython==8.38.0
isoduration==20.11.0
itsdangerous==2.1.2
Janome==0.5.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.5.3
Js2Py==0.71
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
kombu==5.6.2
langcodes==3.5.1
lark==1.3.1
markdown-it-py==4.0.0
markdown2==2.5.3
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib==3.8.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mssql-django==1.4
multidict==6.0.4
multiprocess==0.70.16
multitasking==0.0.12
murmurhash==1.0.15
narwhals==2.13.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook_shim==0.2.4
num2words==0.5.14
numpy==1.26.4
openai==0.27.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.5
overrides==7.7.0
packaging==26.0
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.6
pdf2image==1.17.0
peewee==3.19.0
Pillow==9.1.1
Pint==0.19.2
pipenv==2026.0.3
pipwin==0.5.2
platformdirs==4.5.0
playsound==1.2.2
plotly==6.5.2
plyer==2.1.0
preshed==3.0.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.1
pure_eval==0.2.3
pvporcupine==1.9.5
pyarrow==22.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyjsparser==2.7.1
PyJWT==2.11.0
pymongo==3.12.3
PyMsgBox==1.0.9
pyngrok==7.5.0
pyodbc==5.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==4.0.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==304
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
razorpay==2.0.0
redis==7.2.0
referencing==0.37.0
regex==2022.4.24
reportlab==4.4.10
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9
safetensors==0.7.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
Send2Trash==2.1.0
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.3.2.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
sqlparse==0.5.4
srsly==2.5.2
stack-data==0.6.3
starlette==0.35.1
streamlit==1.52.1
streamlit-lottie==0.0.5
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
thinc==8.2.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.4.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.5.3
tqdm==4.64.0
traitlets==5.14.3
typer==0.9.4
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.3
tzlocal==4.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.27.0
vine==5.1.0
virtualenv==20.35.4
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.3.4
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==2.1.2
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==2.1.1
xxhash==3.6.0
yarl==1.8.2
yfinance==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0


# IndicTrans2 Translation Dependencies
transformers==4.41.0
sentencepiece==0.2.0
sacremoses==0.1.1

